requirements.txt
flask
anthropic
google-cloud-secret-manager
replicate